refactor: Flatten module structure to follow Python standards
- Reorganize project from src/backend/ to src/math_agent/
- Create clean module hierarchy with api/, core/, services/, storage/
- Split monolithic server.py into modular route files
- Update all imports to use new structure
- Fix deprecation warning by removing @app.on_event
- Add proper __init__.py files for all packages
- Update tests to reflect new import paths

The new structure eliminates awkward relative imports and provides
better separation of concerns.
