--- src/backend/__init__.py
@@ -1,1 +1,0 @@
-# Math Agent Backend Package
