+++ src/math_agent/__init__.py
@@ -1,0 +1,2 @@
+"""Math Agent - A benchmarking system for evaluating AI agents on mathematical problems"""
+__version__ = "0.1.0"
