+++ src/math_agent/api/__init__.py
@@ -1,0 +1,1 @@
+"""API module for the math agent system"""
