+++ src/math_agent/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core domain logic and models for the math agent system"""
