+++ src/math_agent/core/models.py
