+++ src/math_agent/core/utils.py
