+++ src/math_agent/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service layer for the math agent system"""
