+++ src/math_agent/storage/__init__.py
@@ -1,0 +1,1 @@
+"""Storage layer for the math agent system"""
